+++ spade_anomaly_detection/parameters_test.py
@@ -49,6 +49,7 @@ class ParametersTest(tf.test.TestCase):
             positive_data_value='1',
             negative_data_value='0',
             unlabeled_data_value='-1',
+            voting_strategy=parameters.VotingStrategy.UNANIMOUS,
         )
     with self.subTest(name='no_input_sources_specified'):
       with self.assertRaises(ValueError):
@@ -61,6 +62,7 @@ class ParametersTest(tf.test.TestCase):
             positive_data_value='1',
             negative_data_value='0',
             unlabeled_data_value='-1',
+            voting_strategy=parameters.VotingStrategy.UNANIMOUS,
         )
 
   def test_equal_data_value_parameter_raises(self):
@@ -88,7 +90,26 @@ class ParametersTest(tf.test.TestCase):
           positive_data_value=1,
           negative_data_value=0,
           unlabeled_data_value=-1,
+          voting_strategy=parameters.VotingStrategy.UNANIMOUS,
       )
+
+  def test_alpha_negative_pseudolabels_not_set_reset_to_alpha(self):
+    parameters_object = parameters.RunnerParameters(
+        train_setting=parameters.TrainSetting.PNU,
+        input_bigquery_table_path='some_project.some_dataset.some_table',
+        data_input_gcs_uri=None,
+        output_gcs_uri='gs://some_bucket/some_path',
+        label_col_name='y',
+        labels_are_strings=False,
+        positive_data_value=1,
+        negative_data_value=0,
+        unlabeled_data_value=-1,
+        alpha=0.8,
+        alpha_negative_pseudolabels=None,
+        voting_strategy=parameters.VotingStrategy.UNANIMOUS,
+    )
+    self.assertEqual(parameters_object.alpha, 0.8)
+    self.assertEqual(parameters_object.alpha_negative_pseudolabels, 0.8)
 
 
 if __name__ == '__main__':