+++ CHANGELOG.md
@@ -24,6 +24,15 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-01-22
+
+* Major revision to extend SPADE with new capabilities. Now it is possible to
+set a voting strategy for the pseudolabeler. It is possible to have a separate
+number of GMM components per model. The `alpha` weight parameter can now be set
+separately for positive and negative pseudolabels.
+* Allow labels to be arbitrary strings.
+* Upgrade Docker base image to Tensorflow 2.17.
+
 ## [0.3.3] - 2024-08-05
 
 * Add support for wildcards in GCS URIs in CSV data loader.
@@ -58,7 +67,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 * Initial release
 
-[Unreleased]: https://github.com/google-research/spade_anomaly_detection/compare/v0.3.3...HEAD
+[Unreleased]: https://github.com/google-research/spade_anomaly_detection/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/google-research/spade_anomaly_detection/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/google-research/spade_anomaly_detection/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/google-research/spade_anomaly_detection/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/google-research/spade_anomaly_detection/compare/v0.3.0...v0.3.1